+++ SpeedyWeather/test/GPU/runtests.jl
@@ -55,8 +55,12 @@ include("primitive_wet.jl")
 if gpu_backend === :CUDA
 
     include("CUDA/architecture.jl")
-    # REACTANT ON GPU (currently only tested with CUDA)
-    include("reactant.jl")
+
+    # CUDA-GRAPHS ACCELERATED FOURIER TRANSFORM (CUDA-only feature)
+    include("cuda_graphs.jl")
+
+    # REACTANT ON GPU (currently only works with CUDA)
+    #include("reactant.jl")
 
 elseif gpu_backend === :AMDGPU